Relocate `get_device_type_from_data` to `simplipy.device`
